Clean HTML
Very basic method to remove tags (comments, start tags and end tags).
Their content is preserved for now (unless for comments, of course).
Added a simple test file to test if tags are removed, and if
other less than are preserved. Hopefully this would work for most cases.
In the future a table of tags and behaviours might be relevant (for each
tag, to know if its content should be kept or removed, etc.)